+++ crates/aiken-lang/src/parser/expr/assignment.rs
@@ -9,25 +9,7 @@ use crate::{
 pub fn let_(
     r: Recursive<'_, Token, UntypedExpr, ParseError>,
 ) -> impl Parser<Token, UntypedExpr, Error = ParseError> + '_ {
-    assignment(r, ast::AssignmentKind::Let)
-}
-
-pub fn expect(
-    r: Recursive<'_, Token, UntypedExpr, ParseError>,
-) -> impl Parser<Token, UntypedExpr, Error = ParseError> + '_ {
-    assignment(r, ast::AssignmentKind::Expect)
-}
-
-fn assignment(
-    r: Recursive<'_, Token, UntypedExpr, ParseError>,
-    kind: ast::AssignmentKind,
-) -> impl Parser<Token, UntypedExpr, Error = ParseError> + '_ {
-    let keyword = match kind {
-        ast::AssignmentKind::Let => Token::Let,
-        ast::AssignmentKind::Expect => Token::Expect,
-    };
-
-    just(keyword)
+    just(Token::Let)
         .ignore_then(pattern())
         .then(just(Token::Colon).ignore_then(annotation()).or_not())
         .then_ignore(just(Token::Equal))
@@ -37,10 +19,48 @@ fn assignment(
                 location: span,
                 value: Box::new(value),
                 pattern,
-                kind,
+                kind: ast::AssignmentKind::Let,
                 annotation,
             },
         )
+}
+
+pub fn expect(
+    r: Recursive<'_, Token, UntypedExpr, ParseError>,
+) -> impl Parser<Token, UntypedExpr, Error = ParseError> + '_ {
+    just(Token::Expect)
+        .ignore_then(
+            pattern()
+                .then(just(Token::Colon).ignore_then(annotation()).or_not())
+                .then_ignore(just(Token::Equal))
+                .or_not(),
+        )
+        .then(r.clone())
+        .map_with_span(move |(opt_pattern, value), span| {
+            let (pattern, annotation) = opt_pattern.unwrap_or_else(|| {
+                (
+                    ast::UntypedPattern::Constructor {
+                        is_record: false,
+                        location: span,
+                        name: "True".to_string(),
+                        arguments: vec![],
+                        module: None,
+                        constructor: (),
+                        with_spread: false,
+                        tipo: (),
+                    },
+                    None,
+                )
+            });
+
+            UntypedExpr::Assignment {
+                location: span,
+                value: Box::new(value),
+                pattern,
+                kind: ast::AssignmentKind::Expect,
+                annotation,
+            }
+        })
 }
 
 #[cfg(test)]
@@ -55,5 +75,10 @@ mod tests {
     #[test]
     fn expect() {
         assert_expr!("expect Some(x) = something.field");
+    }
+
+    #[test]
+    fn expect_bool_sugar() {
+        assert_expr!("expect something.field == wow");
     }
 }