+++ crates/aiken-lang/src/parser/utils.rs
@@ -62,6 +62,27 @@ macro_rules! assert_annotation {
 }
 
 #[macro_export]
+macro_rules! assert_pattern {
+    ($code:expr) => {
+        use chumsky::Parser;
+
+        let $crate::parser::lexer::LexInfo { tokens, .. } = $crate::parser::lexer::run(indoc::indoc! { $code }).unwrap();
+
+        let stream = chumsky::Stream::from_iter($crate::ast::Span::create(tokens.len(), 1), tokens.into_iter());
+
+        let result = $crate::parser::pattern().parse(stream).unwrap();
+
+        insta::with_settings!({
+            description => concat!("Code:\n\n", indoc::indoc! { $code }),
+            prepend_module_to_snapshot => false,
+            omit_expression => true
+        }, {
+            insta::assert_debug_snapshot!(result);
+        });
+    };
+}
+
+#[macro_export]
 macro_rules! assert_module {
     ($code:expr) => {
         let (module, _) =